internal: scalable module structure for fixits
